refactor: extract assignment, tickets, issues, whats_new, trash into per-app packages
Fourth phase of the restructure — assignment goes first (full extraction,
since app/content/ and app/courses/ in later phases will need it to
already exist as its own app), then a mechanical sweep of four small,
low-fan-in blueprints using the now-proven recipe:

- app/models/assignment.py + app/blueprints/assignment.py -> app/assignment/
  (hoisted 2 duplicate local authorization imports to the top)
- app/blueprints/tickets.py + app/services/ticket_service.py -> app/tickets/
  (superadmin.py and test_multitenant.py, which also use ticket_service,
  updated to the new import path; email_service.py stays shared)
- app/blueprints/issues.py -> app/issues/. Its /api url_prefix meant the
  page-render route (previously living in main.py as `/issues`) couldn't
  join the same blueprint, so it now lives on a second, prefix-free
  `issues_pages_bp` in the same module — the Flask analog of Django apps
  splitting views.py per concern. Updated the 4 template call sites
  (base.html x2, course_detail.html, _sidebar.html) that referenced the
  old `main.issues` endpoint to `issues_pages.issues_page`.
- app/models/whats_new.py + app/blueprints/whats_new_view.py -> app/whats_new/
- app/blueprints/trash.py -> app/trash/ (the "Ruang TPS" soft-delete
  recycle bin; touches Course/Quiz/Assignment/File/Link/ContentFolder
  directly since that's its actual job, not a boundary violation)

All templates for these features moved into app/<feature>/templates/<feature>/
with matching Blueprint(template_folder=...) declarations and updated
render_template() paths. app/models/__init__.py updated for assignment
and whats_new; ticket/issue models stay in app/models/ per the plan
(genuinely cross-cutting, not worth relocating for two model files).

Verified: create_app() builds with all 224 routes intact (net zero: +1 for
issues_pages_bp, -1 removed from main.py), SQLAlchemy metadata still has
33 tables, and pytest is unchanged from baseline (25 pre-existing
failures, 99 passing).
