+++ app/assignment/templates/assignment/assignment_detail.html
