+++ app/templates/_sidebar.html
@@ -110,7 +110,7 @@
         <!-- Bantuan & Masalah -->
         <div class="pt-6">
             <p class="px-4 pb-2 text-[11px] font-black text-[#afafaf] uppercase tracking-[0.15em] sidebar-hide">Bantuan</p>
-            <a href="{{ url_for('main.issues') }}" class="nav-link-duo {% if request.endpoint == 'main.issues' %}active{% endif %}">
+            <a href="{{ url_for('issues_pages.issues_page') }}" class="nav-link-duo {% if request.endpoint == 'issues_pages.issues_page' %}active{% endif %}">
                 <div class="nav-icon-duo bg-[#ff4b4b]">
                     <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2.5" d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"/>
