+++ app/templates/base.html
@@ -114,13 +114,13 @@
                 <span class="text-[10px] font-black uppercase tracking-widest {{ 'text-primary-600' if request.endpoint == 'main.dashboard' else 'text-gray-400' }}">Home</span>
             </a>
 
-            <a href="{{ url_for('main.issues') }}" class="flex flex-col items-center space-y-1 group">
-                <div class="p-2 rounded-2xl transition-all {{ 'bg-primary-50 text-primary-600' if request.endpoint == 'main.issues' else 'text-gray-400 group-hover:bg-gray-50 dark:group-hover:bg-gray-800' }}">
+            <a href="{{ url_for('issues_pages.issues_page') }}" class="flex flex-col items-center space-y-1 group">
+                <div class="p-2 rounded-2xl transition-all {{ 'bg-primary-50 text-primary-600' if request.endpoint == 'issues_pages.issues_page' else 'text-gray-400 group-hover:bg-gray-50 dark:group-hover:bg-gray-800' }}">
                     <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"/>
                     </svg>
                 </div>
-                <span class="text-[10px] font-black uppercase tracking-widest {{ 'text-primary-600' if request.endpoint == 'main.issues' else 'text-gray-400' }}">Issues</span>
+                <span class="text-[10px] font-black uppercase tracking-widest {{ 'text-primary-600' if request.endpoint == 'issues_pages.issues_page' else 'text-gray-400' }}">Issues</span>
             </a>
 
             <a href="{{ url_for('main.settings') }}" class="flex flex-col items-center space-y-1 group">
