+++ app/templates/course_detail.html
@@ -61,7 +61,7 @@
                     <span class="group-hover:text-primary-700 dark:group-hover:text-primary-400 transition-colors">EDIT KELAS</span>
                 </button>
                 {% endif %}
-                <a href="{{ url_for('main.issues') }}" class="p-2 text-primary-600 hover:bg-primary-50 rounded-lg transition-colors" title="Manajemen Masalah">
+                <a href="{{ url_for('issues_pages.issues_page') }}" class="p-2 text-primary-600 hover:bg-primary-50 rounded-lg transition-colors" title="Manajemen Masalah">
                     <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"/>
                     </svg>
