+++ app/trash/templates/trash/course_tps.html
